Convert rest of codebase to ES6 and move routes to separate folders
