+++ lib/auth.js
@@ -1,25 +1,27 @@
-var debug = require('debug')('SIR:auth');
-var passport = require('passport');
-var GitHubStrategy = require('passport-github2').Strategy;
-var express = require('express');
-var router = express.Router();
+import bug from 'debug';
+const debug = bug('SIR:auth');
+import passport from 'passport';
+import githubPassport from 'passport-github2';
+const GitHubStrategy = githubPassport.Strategy;
+import express from 'express';
+const router = express.Router();
 
-var clientId = process.env.CLIENT_ID;
-var clientSecret = process.env.CLIENT_SECRET;
-var nodeEnv = process.env.NODE;
-var baseUrl = nodeEnv === 'production' ?
+const clientId = process.env.CLIENT_ID;
+const clientSecret = process.env.CLIENT_SECRET;
+const nodeEnv = process.env.NODE;
+const baseUrl = nodeEnv === 'production' ?
   process.env.BASE_URL : 'http://localhost:5000';
 
 debug('client id', clientId)
 debug('client secret', clientSecret)
 debug(baseUrl + '/auth/github/callback')
 
-passport.serializeUser(function (user, done) {
+passport.serializeUser((user, done) => {
   debug('serialize', user)
   done(null, user.id);
 });
 
-passport.deserializeUser(function (user, done) {
+passport.deserializeUser((user, done) => {
   debug('deserialize', user)
   done(null, user);
 });
@@ -28,7 +30,7 @@ passport.use(new GitHubStrategy({
   clientID: clientId,
   clientSecret: clientSecret,
   callbackURL: baseUrl + '/auth/github/callback'
-}, function(accessToken, refreshToken, profile, done) {
+}, (accessToken, refreshToken, profile, done) => {
   // debug('auth', profile)
   return done(null, profile);
 }));
@@ -40,4 +42,4 @@ router.get('/github/callback', passport.authenticate('github', {
   failureRedirect: '/signin'
 }))
 
-module.exports = router;
+export default router;
