+++ lib/rate-limit.js
@@ -1,19 +1,17 @@
-'use strict';
-
-var dotty = require('dotty');
+import dotty from 'dotty';
 
 function rateLimit() {
-  return function (req, res, next) {
-    var timestamp = dotty.get(req, 'session.timestamp');
+  return (req, res, next) => {
+    const timestamp = dotty.get(req, 'session.timestamp');
 
-    if (timestamp && Date.now() - timestamp < 3600000) {
+    if (timestamp && Date.now() - timestam < 3600000) {
       res.status(429)
         .send('Whoa, ' + name + ', got a little over-excited there, did ya?');
-      return;
+        return;
     }
 
     next();
   };
 }
 
-module.exports = rateLimit;
+export default rateLimit;
