+++ lib/slack.js
@@ -1,6 +1,4 @@
-'use strict';
-
-var got = require('got');
+import got from 'got';
 
 function slack(url) {
   return function send(payload) {
@@ -11,4 +9,4 @@ function slack(url) {
   };
 }
 
-module.exports = slack;
+export default slack;
