+++ lib/validate.js
@@ -1,10 +1,8 @@
-'use strict';
-
-var dotty = require('dotty');
+import dotty from 'dotty';
 
 function validateLogin() {
-  return function (req, res, next) {
-    var user = dotty.get(req, 'session.user');
+  return (req, res, next) => {
+    const user = dotty.get(req, 'session.user');
 
     if (!user) {
       req.session.error = 'Not Authenticated';
@@ -16,4 +14,4 @@ function validateLogin() {
   };
 }
 
-module.exports = validateLogin;
+export default validateLogin;
