+++ public/scripts/main.js
@@ -1,9 +1,9 @@
 (function () {
-  var po = document.createElement('script');
+  const po = document.createElement('script');
   po.type = 'text/javascript';
   po.async = true;
   po.src = 'https://plus.google.com/js/client:plusone.js?onload=start';
-  var s = document.getElementsByTagName('script')[0];
+  const s = document.getElementsByTagName('script')[0];
   s.parentNode.insertBefore(po, s);
 }());
 
@@ -11,14 +11,14 @@ function signInCallback(authResult) {
   if (authResult.code) {
     $.ajax({
       url: 'https://www.googleapis.com/plus/v1/people/me',
-      beforeSend: function (xhr) {
+      beforeSend: (xhr) => {
         xhr.setRequestHeader('Authorization','Bearer ' + authResult.access_token);
       }
-    }).done(function (user) {
+    }).done((user) => {
       $.post('/signin', {
         code: authResult.code,
         user: user
-      }).done(function () {
+      }).done(() => {
         window.location.replace('/apply')
       });
     });
