+++ server.js
@@ -31,10 +31,7 @@ const channel = env.SLACK_CHANNEL;
 const botName = env.SLACK_BOT_NAME || 'SIR';
 const nodeEnv = env.NODE;
 
-function exitWithError(err) {
-  error(err);
-  process.exit(1);
-}
+import { exitWithError } from './lib/helper-functions';
 
 if (!gaToken) {
   exitWithError('Please set GA_TOKEN environment variable.')
@@ -54,29 +51,6 @@ if (!env.CLIENT_SECRET) {
 
 import mod from './lib/slack';
 const slack = mod(slackUrl);
-
-function getStrings() {
-  const strings = yaml.safeLoad(fs.readFileSync(path.resolve('./strings.yml')));
-
-  // extend strings
-  strings.main = _.assign({}, {
-    title: strings.title,
-    gaToken: gaToken
-  }, strings.main);
-
-  strings.signin = _.assign({}, {
-    title: strings.title,
-    gaToken: gaToken,
-    // clientId: clientId
-  }, strings.signin);
-
-  strings.apply = _.assign({}, {
-    title: strings.title,
-    gaToken: gaToken
-  }, strings.apply);
-
-  return strings;
-}
 
 app.use(logger('dev'));
 
@@ -106,101 +80,17 @@ app.use(
   }
 );
 
+import apply from './routes/apply';
+import home from './routes/home';
+import signin from './routes/signin';
+import thanks from './routes/thanks';
+app.use('/apply', apply);
+app.use('/', home);
+app.use('/signin', signin);
+app.use('/thanks', thanks);
+
 import auth from './lib/auth';
 app.use('/auth', auth);
-
-app.get('/', (req, res) => {
-  res.render('main', _.assign({}, getStrings().main, dotty.get(req, 'session.user')));
-});
-
-app.get('/signin', (req, res) => {
-  const user = dotty.get(req, 'session.user');
-
-  if (user) {
-    res.redirect('/apply');
-  } else {
-    res.render('signin', getStrings().signin);
-  }
-});
-
-app.get('/thanks', validate(), (req, res) => {
-  res.render('thanks', _.assign({}, getStrings().main, dotty.get(req, 'session.user')));
-});
-
-app.get('/apply', validate(), (req, res) => {
-  const user = dotty.get(req, 'session.user');
-  const strings = getStrings();
-
-  strings.apply.form.fullName.value = user.displayName;
-  strings.apply.form.email.value = user.emails[0].value;
-
-  res.render('apply', _.assign({}, strings.apply, user));
-});
-
-app.post('/apply', validate(), rateLimit(), (req, res) => {
-  const user = dotty.get(req, 'session.user');
-  const files = req.files;
-  const renameJobs = [];
-
-  debug('Received application from "%s <%s>"', user.displayName, user.emails[0].value);
-
-  for (const field in files) {
-    const fileObj = files[field];
-    const tmpPath = fileObj.path;
-    const filename = field + '-' + fileObj.name;
-    const dest = __dirname + '/public/images/' + filename;
-
-    _.assign(fileObj, {
-      dest: dest,
-      uri: req.originUri + '/images/' + filename
-    });
-
-    renameJobs.push(async.apply(mv, tmpPath, dest));
-
-    async.parallel(renameJobs, (err) => {
-      if (err) {
-        error(err);
-        return res.sendStatus(500);
-      }
-
-      res.redirect('/thanks');
-
-      slack({
-      channel: channel,
-      username: botName,
-      icon_url: req.originUri + 'images/bot.png',
-      attachments: [
-        {
-          fallback: user.displayName + ' wants to join Slack',
-          author_name: user.displayName,
-          author_link: user.url,
-          author_icon: dotty.get(user, 'image.url'),
-          color: '#28f428',
-          pretext: 'New invite request:',
-          text: req.body.comments ? '"' + req.body.comments + '"' : undefined,
-          fields: _.map(
-            _.pairs(_.omit(req.body, 'comments')),
-            _.flow(
-              x => x,
-              _.partialRight(_.map, (str, i) => {
-                return i ? str : changeCase.title(str);
-              }),
-              _.partial(_.zipObject, ['title', 'value']),
-              _.partialRight(_.assign, { short: true })
-            )
-          )
-            .concat(_.map(files, (file) => (
-              {
-                title: file.fieldname,
-                value: '<' + file.uri + '|View>',
-                short: true
-              }
-            )))
-        }
-      ]
-    });
-  });
-}});
 
 app.use((err, req, res, next) => {
   error(err.stack);
